Move demos to packages, fix errors and lint.
