+++ demos/index.html
@@ -1,4 +1,7 @@
+<h1>Glimmer Demos</h1>
+<div><a href="./ripples.html">Ripples</a></div>
+<div><a href="./uptime-boxes.html">Uptime Boxes</a></div>
+<div><a href="./conways.html">Conways</a></div>
+<div><a href="./visualizer.html">Glimmer Visualizer</a></div>
 
-<a href="./ripples.html">Ripples</a>
-<a href="./uptime-boxes.html">Uptime Boxes</a>
-<a href="./visualizer.html">Glimmer Visualizer</a>
+
