+++ demos/visualizer.html
@@ -187,11 +187,11 @@
     <script src="../amd/glimmer-common.amd.js"></script>
     <script src="../amd/glimmer-compiler.amd.js"></script>
     <script src="../amd/glimmer-runtime.amd.js"></script>
-    <script src="./demos.amd.js"></script>
+    <script src="../amd/glimmer-demos.amd.js"></script>
   </head>
   <body>
     <script type="application/javascript">
-      require("glimmer-demos/visualizer").init();
+      require("glimmer-demos").Visualizer.init();
     </script>
   </body>
 </html>
