+++ packages/glimmer-demos/index.ts
@@ -1,0 +1,10 @@
+import * as uptime from "./lib/uptime";
+import * as conways from "./lib/conways";
+import * as ripples from "./lib/ripples";
+import * as visualizer from "./lib/visualizer";
+import * as stats from "./lib/stats";
+export const UptimeDemo = uptime;
+export const RipplesDemo = ripples;
+export const ConwaysDemo = conways;
+export const Visualizer = visualizer;
+export const Stats = stats; // TODO does this belong in demos?
